# spec/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Business Watchdog core API
  version: 1.0.0
  description: |
    Core contracts for P0 and P1. This document is a specification, not a running API.
    HMAC canonicalization and transport limits are normative in TZ sections 11 and 27.
    UI CRUD, auth, billing, export and credential rotation must be added before their implementation.
    All monetary amounts are decimal strings in minor units. All timestamps are RFC3339 UTC.
servers:
  - url: https://watchdog.example.invalid
...
paths:
...
  /api/v1/tenants/context:
    get:
      tags: [Tenancy]
      operationId: tenantContext
      security:
        - {UserSession: []}
      responses:
        "200":
          description: Resolved tenant context
          content:
            application/json:
              schema:
                type: object
                properties:
                  active_tenant_id: {type: [string, "null"], format: uuid}
                  tenant_context_id: {type: [string, "null"], format: uuid}
                  tenant_context_source: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
...
  /api/v1/incidents:
    get:
      tags: [Incidents]
      operationId: listIncidents
      security:
        - {UserSession: []}
      parameters:
        - name: store_id
          in: query
          schema: {type: string, format: uuid}
        - name: state
          in: query
          schema: {enum: [open, acknowledged, resolved]}
        - name: severity
          in: query
          schema: {enum: [info, warning, critical]}
        - name: family
          in: query
          schema: {type: string}
        - name: currency
          in: query
          schema: {type: string, pattern: "^[A-Z]{3}$"}
        - name: limit
          in: query
          schema: {type: integer, minimum: 1, maximum: 100}
        - name: cursor
          in: query
          schema: {type: string, maxLength: 2048}
      responses:
        "200":
          description: Tenant-scoped stable cursor page
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: {$ref: "#/components/schemas/Incident"}
                  next_cursor: {type: [string, "null"]}
        "401": {$ref: "#/components/responses/Problem"}
  /api/v1/incidents/{incident_id}:
    get:
      tags: [Incidents]
      operationId: getIncident
      security:
        - {UserSession: []}
      parameters:
        - name: incident_id
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        "200":
          description: Incident with signals and activity
          content:
            application/json:
              schema: {$ref: "#/components/schemas/IncidentDetail"}
          headers:
            ETag:
              description: Quoted version for If-Match
              schema: {type: string}
        "401": {$ref: "#/components/responses/Problem"}
        "404": {$ref: "#/components/responses/Problem"}
...
components:
  securitySchemes:
    IntegrationKeyId:
      type: apiKey
      in: header
      name: X-BW-Key-Id
      description: Key ID plus mandatory raw-body HMAC signature headers; key ID alone never authenticates.
    UserSession:
      type: apiKey
      in: cookie
      name: watchdog_session
      description: Same-origin Sanctum session with CSRF on mutations.
    WorkerToken:
      type: http
      scheme: bearer
      description: Scoped worker credential restricted to private ingress.
...
  responses:
    Problem:
      description: Structured safe problem response; HTTP status determines retry policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
    MoneyMinor:
      type: string
      pattern: '^-?(0|[1-9][0-9]{0,18})$'
      description: Signed 64-bit integer encoded as decimal string; semantic validator enforces bigint bounds.
    Problem:
      type: object
      additionalProperties: false
      required: [code, message, request_id]
      properties:
        code: {type: string}
        message: {type: string}
        request_id: {type: string, format: uuid}
        errors: {type: array, items: {type: object}}
        retry_after_seconds: {type: integer, minimum: 0}
...
    Incident:
      type: object
      required:
        - id
        - store_id
        - family
        - component
        - state
        - severity
        - title_code
        - revision
        - first_seen_at
        - last_seen_at
      properties:
        id: {type: string, format: uuid}
        store_id: {type: string, format: uuid}
        family: {enum: [money, checkout, checkout_payment, integration]}
        component: {type: string}
        state: {enum: [open, acknowledged, resolved]}
        severity: {enum: [info, warning, critical]}
        title_code: {type: string, description: Stable code; the UI translates it (ru/en/de).}
        currency: {type: [string, "null"], pattern: "^[A-Z]{3}$"}
        verified_discrepancy_minor:
          oneOf:
            - {$ref: "#/components/schemas/MoneyMinor"}
            - {type: "null"}
        first_seen_at: {type: string, format: date-time}
        last_seen_at: {type: string, format: date-time}
        last_good_at: {type: [string, "null"], format: date-time}
        first_bad_at: {type: [string, "null"], format: date-time}
        acknowledged_by: {type: [string, "null"], format: uuid}
        acknowledged_at: {type: [string, "null"], format: date-time}
        resolved_at: {type: [string, "null"], format: date-time}
        resolution_reason: {type: [string, "null"]}
        revision: {type: integer, minimum: 1}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    Signal:
      type: object
      required: [id, signal_type, family, severity, confidence, detected_at]
      properties:
        id: {type: string, format: uuid}
        signal_type: {enum: [reconciliation_finding, payment_attempts, connector_freshness, browser_check]}
        family: {type: string}
        component: {type: string}
        severity: {enum: [info, warning, critical]}
        confidence: {enum: [observed, corroborated, inferred, unknown]}
        rule_version: {type: string}
        config_version: {type: integer}
        currency: {type: [string, "null"], pattern: "^[A-Z]{3}$"}
        finding_id: {type: [string, "null"], format: uuid}
        evidence: {type: object, description: Sanitized facts; no customer data.}
        detected_at: {type: string, format: date-time}
    IncidentActivity:
      type: object
      required: [id, incident_id, kind, created_at]
      properties:
        id: {type: string, format: uuid}
        incident_id: {type: string, format: uuid}
        kind:
          enum: [created, signal_linked, acknowledged, comment, resolved, reopened, severity_changed, suppressed]
        actor_id: {type: [string, "null"], format: uuid}
        incident_revision: {type: integer}
        data: {type: object}
        created_at: {type: string, format: date-time}
    IncidentDetail:
      allOf:
        - {$ref: "#/components/schemas/Incident"}
        - type: object
          required: [signals, activity]
          properties:
            signals:
              type: array
              items: {$ref: "#/components/schemas/Signal"}
            activity:
              type: array
              items: {$ref: "#/components/schemas/IncidentActivity"}
...